refactor(P4-B): split analytics helpers and isolate DB writes for interaction & session analytics services
- Added TodayFeedAnalyticsRemoteDataSource to centralize Supabase writes
- Extracted analytics and session data models into domain models package
- Removed in-file model definitions to bring services below size ceiling
- Updated services to delegate DB writes via remote datasource
- Adjusted dashboard models import
- Marked P4-B status ✅ in tech plan
