+++ app/lib/features/today_feed/data/services/session_duration_tracking_service.dart
@@ -6,7 +6,8 @@ import 'package:flutter/foundation.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import '../../domain/models/today_feed_content.dart';
 import '../../../../core/services/connectivity_service.dart';
-import '../../../../core/services/version_service.dart';
+import '../datasources/today_feed_analytics_remote_datasource.dart';
+import '../../domain/models/reading_session_models.dart';
 
 /// Configuration constants for session duration tracking
 class SessionTrackingConfig {
@@ -30,169 +31,6 @@ class SessionTrackingConfig {
   static const double highEngagementRate = 0.8; // 80% of estimated reading time
 }
 
-/// Session quality metrics for engagement analysis
-enum SessionQuality {
-  brief('brief'), // < 30 seconds
-  moderate('moderate'), // 30s - 2 minutes
-  engaged('engaged'), // 2-5 minutes
-  deep('deep'); // > 5 minutes
-
-  const SessionQuality(this.value);
-  final String value;
-
-  static SessionQuality fromDuration(Duration duration) {
-    if (duration < SessionTrackingConfig.shortReadThreshold) {
-      return SessionQuality.brief;
-    } else if (duration < SessionTrackingConfig.mediumReadThreshold) {
-      return SessionQuality.moderate;
-    } else if (duration < SessionTrackingConfig.longReadThreshold) {
-      return SessionQuality.engaged;
-    } else {
-      return SessionQuality.deep;
-    }
-  }
-}
-
-/// Reading session data model for analytics
-class ReadingSession {
-  final String sessionId;
-  final String userId;
-  final int contentId;
-  final DateTime startTime;
-  final DateTime endTime;
-  final Duration duration;
-  final SessionQuality quality;
-  final int samplesCount;
-  final double engagementScore;
-  final Map<String, dynamic> metadata;
-
-  const ReadingSession({
-    required this.sessionId,
-    required this.userId,
-    required this.contentId,
-    required this.startTime,
-    required this.endTime,
-    required this.duration,
-    required this.quality,
-    required this.samplesCount,
-    required this.engagementScore,
-    required this.metadata,
-  });
-
-  /// Create reading session from tracking data
-  factory ReadingSession.fromTrackingData({
-    required String sessionId,
-    required String userId,
-    required int contentId,
-    required DateTime startTime,
-    required DateTime endTime,
-    required List<DateTime> activitySamples,
-    required TodayFeedContent content,
-    Map<String, dynamic>? additionalMetadata,
-  }) {
-    final duration = endTime.difference(startTime);
-    final quality = SessionQuality.fromDuration(duration);
-    final engagementScore = _calculateEngagementScore(
-      duration,
-      content.estimatedReadingMinutes,
-      activitySamples.length,
-    );
-
-    return ReadingSession(
-      sessionId: sessionId,
-      userId: userId,
-      contentId: contentId,
-      startTime: startTime,
-      endTime: endTime,
-      duration: duration,
-      quality: quality,
-      samplesCount: activitySamples.length,
-      engagementScore: engagementScore,
-      metadata: {
-        'content_title': content.title,
-        'content_category': content.topicCategory.value,
-        'estimated_reading_minutes': content.estimatedReadingMinutes,
-        'ai_confidence_score': content.aiConfidenceScore,
-        'platform': defaultTargetPlatform.name,
-        'app_version': VersionService.appVersion,
-        ...?additionalMetadata,
-      },
-    );
-  }
-
-  /// Calculate engagement score based on duration and content
-  static double _calculateEngagementScore(
-    Duration actualDuration,
-    int estimatedMinutes,
-    int samplesCount,
-  ) {
-    if (estimatedMinutes <= 0) return 0.0;
-
-    final estimatedDuration = Duration(minutes: estimatedMinutes);
-    final baseScore = math.min(
-      actualDuration.inMilliseconds / estimatedDuration.inMilliseconds,
-      2.0, // Cap at 2x estimated time
-    );
-
-    // Adjust for sampling quality (more samples = higher confidence)
-    final samplingQuality = math.min(samplesCount / 10.0, 1.0);
-
-    return (baseScore * samplingQuality).clamp(0.0, 2.0);
-  }
-
-  /// Convert to JSON for database storage
-  Map<String, dynamic> toJson() {
-    return {
-      'session_id': sessionId,
-      'user_id': userId,
-      'content_id': contentId,
-      'start_time': startTime.toIso8601String(),
-      'end_time': endTime.toIso8601String(),
-      'duration_seconds': duration.inSeconds,
-      'session_quality': quality.value,
-      'samples_count': samplesCount,
-      'engagement_score': engagementScore,
-      'metadata': metadata,
-    };
-  }
-
-  /// Create from JSON
-  factory ReadingSession.fromJson(Map<String, dynamic> json) {
-    return ReadingSession(
-      sessionId: json['session_id'] as String,
-      userId: json['user_id'] as String,
-      contentId: json['content_id'] as int,
-      startTime: DateTime.parse(json['start_time'] as String),
-      endTime: DateTime.parse(json['end_time'] as String),
-      duration: Duration(seconds: json['duration_seconds'] as int),
-      quality: SessionQuality.values.firstWhere(
-        (q) => q.value == json['session_quality'],
-        orElse: () => SessionQuality.brief,
-      ),
-      samplesCount: json['samples_count'] as int,
-      engagementScore: (json['engagement_score'] as num).toDouble(),
-      metadata: json['metadata'] as Map<String, dynamic>,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-    return other is ReadingSession &&
-        other.sessionId == sessionId &&
-        other.userId == userId &&
-        other.contentId == contentId;
-  }
-
-  @override
-  int get hashCode => Object.hash(sessionId, userId, contentId);
-
-  @override
-  String toString() {
-    return 'ReadingSession(id: $sessionId, duration: ${duration.inSeconds}s, quality: ${quality.value})';
-  }
-}
-
 /// Active session tracker for real-time duration monitoring
 class _ActiveSessionTracker {
   final String sessionId;
@@ -214,18 +52,15 @@ class _ActiveSessionTracker {
     required this.content,
   }) : lastActivity = startTime;
 
-  /// Start periodic activity sampling
   void startSampling() {
     _samplingTimer?.cancel();
     _samplingTimer = Timer.periodic(
       SessionTrackingConfig.samplingInterval,
       (_) => _recordActivitySample(),
     );
-
     _resetTimeoutTimer();
   }
 
-  /// Record user activity sample
   void _recordActivitySample() {
     final now = DateTime.now();
     activitySamples.add(now);
@@ -233,22 +68,18 @@ class _ActiveSessionTracker {
     _resetTimeoutTimer();
   }
 
-  /// Reset session timeout timer
   void _resetTimeoutTimer() {
     _timeoutTimer?.cancel();
     _timeoutTimer = Timer(
       SessionTrackingConfig.sessionTimeout,
-      () => _onSessionTimeout(),
+      _onSessionTimeout,
     );
   }
 
-  /// Handle session timeout
   void _onSessionTimeout() {
     debugPrint('⏰ Session timeout for session $sessionId');
-    // Session will be finalized by the service
   }
 
-  /// Record user interaction (tap, scroll, etc.)
   void recordInteraction() {
     final now = DateTime.now();
     lastActivity = now;
@@ -256,7 +87,6 @@ class _ActiveSessionTracker {
     _resetTimeoutTimer();
   }
 
-  /// Finalize session and create ReadingSession
   ReadingSession finalize({Map<String, dynamic>? additionalMetadata}) {
     _samplingTimer?.cancel();
     _timeoutTimer?.cancel();
@@ -274,68 +104,9 @@ class _ActiveSessionTracker {
     );
   }
 
-  /// Dispose resources
   void dispose() {
     _samplingTimer?.cancel();
     _timeoutTimer?.cancel();
-  }
-}
-
-/// Session analytics aggregation data
-class SessionAnalytics {
-  final int totalSessions;
-  final Duration totalReadingTime;
-  final Duration averageSessionDuration;
-  final double averageEngagementScore;
-  final Map<SessionQuality, int> qualityDistribution;
-  final Map<String, double> topicEngagement;
-  final DateTime? lastSessionTime;
-  final int consecutiveDaysWithSessions;
-
-  const SessionAnalytics({
-    required this.totalSessions,
-    required this.totalReadingTime,
-    required this.averageSessionDuration,
-    required this.averageEngagementScore,
-    required this.qualityDistribution,
-    required this.topicEngagement,
-    this.lastSessionTime,
-    required this.consecutiveDaysWithSessions,
-  });
-
-  factory SessionAnalytics.empty() {
-    return const SessionAnalytics(
-      totalSessions: 0,
-      totalReadingTime: Duration.zero,
-      averageSessionDuration: Duration.zero,
-      averageEngagementScore: 0.0,
-      qualityDistribution: {},
-      topicEngagement: {},
-      consecutiveDaysWithSessions: 0,
-    );
-  }
-
-  /// Calculate reading efficiency score (0.0 - 1.0)
-  double get readingEfficiency {
-    if (totalSessions == 0) return 0.0;
-    return averageEngagementScore.clamp(0.0, 1.0);
-  }
-
-  /// Determine engagement level based on metrics
-  String get engagementLevel {
-    if (averageEngagementScore >= SessionTrackingConfig.highEngagementRate) {
-      return 'high';
-    } else if (averageEngagementScore >=
-        SessionTrackingConfig.minEngagementRate) {
-      return 'medium';
-    } else {
-      return 'low';
-    }
-  }
-
-  @override
-  String toString() {
-    return 'SessionAnalytics(sessions: $totalSessions, avgDuration: ${averageSessionDuration.inSeconds}s, engagement: ${(averageEngagementScore * 100).toStringAsFixed(1)}%)';
   }
 }
 
@@ -356,6 +127,7 @@ class SessionDurationTrackingService {
 
   // Dependencies
   late final SupabaseClient _supabase;
+  late final TodayFeedAnalyticsRemoteDataSource _remote;
   bool _isInitialized = false;
 
   // Active session management
@@ -370,6 +142,7 @@ class SessionDurationTrackingService {
 
     try {
       _supabase = Supabase.instance.client;
+      _remote = TodayFeedAnalyticsRemoteDataSource();
 
       // Set up connectivity monitoring for offline sync
       _connectivitySubscription = ConnectivityService.statusStream.listen(
@@ -619,7 +392,7 @@ class SessionDurationTrackingService {
   /// Sync session data to database
   Future<void> _syncSessionToDatabase(Map<String, dynamic> sessionData) async {
     try {
-      await _supabase.from('today_feed_reading_sessions').insert(sessionData);
+      await _remote.insertReadingSession(sessionData);
       debugPrint('✅ Session synced to database: ${sessionData['session_id']}');
     } catch (e) {
       debugPrint('❌ Failed to sync session to database: $e');
